Lots of little clean-ups... I realized that I started out populating parameter datatypes using the cvmix_put command (from cvmix_*_init routines) but lately have been defining values directly. I like how the code reads with the put commands, though, so I added them everywhere they were needed.
I also made a more substative change in the new double diffusion model. Rrho, the stratification parameter, is a cell-center value, not an interface value. I changed the dimension in the driver file, and then updated the output routines accordingly. If you output to a text file then the "nlev+1" value of Rrho is set to 0, but if you write to a netCDF file then Rrho has dimension nlev rather than nlev+1.