Added physics file to tidal mixing namelist to read in buoyancy frequeny, though currently the file itself is not under version control (it's 50 MB, want to make sure it's in a good format before committing it... it may make sense to merge the physics into the "grid" file, since depth seems to also be a physical quantity). The tidal driver now outputs diffusivity insted of viscosity, since that's what the Simmons scheme sets.
Bigger changes: the io module can now read in a 3D field of doubles from a netcdf file (needed for reading in buoyancy), and the tidal driver has a very different behavior. I'm now interested in the column of ocean at 353.9634 E and 58.84838 N because in POP there seems to be a non-trivial amount of tidal mixing there. For the actual computation, I seem to have a lot to do:

1) send energy flux value to tidal mixing initialization routine
2) compute the vertical deposition function in initialization routine (so we need to send the levels / ocean bottom depth)
3) send buoyancy to tidal mixing computation routine, pull the energy flux and deposition function from initialization to compute diffusivity

Note
--

It probably makes the most sense to store just the time-independent portion of tidal mixing after initialization:

time_independent = mixing_frac*efficiency*energy_flux*vertical_deposition

and divide that by density*buoyancy in the computation stage (POP uses constant density = 1... if that's standard maybe we should move density into time_independent)