I wanted the last build system commit to be the last, but I think this update will let me close out issue 5 (just need to check the cray compiler to make sure upper-case mod files are still handled correctly).
The big change is to makedep - if you give it $(INC_DIR) it will add those files to the dependency file as needed (for example, cvmix_ddiff_drv.o will depend on $(INC_DIR)/cvmix_ddiff.mod) so changes to the library will trigger minimal changes to the standalone executable. I'm 99% sure it'll get $(INC_DIR)/CVMIX_DDIFF.mod correct on the Cray.

Besides saving some time when re-compiling, this also let me remove a pretty big chunk of code from the Makefile. I also tweaked the order that things are done - the library needs to be built before the stand-alone dependencies are made.