Ugh, this is fixing something that shouldn't have been broken in the first place. When I originally wrote the interpolation routine (and the corresponding root-finding code), I assumed the polynomial would be cubic and numbered the coefficients so the equation would be of the form
f(x) = a(1)*x^3 + a(2)*x^2 + a(3)*x + a(4)

(This was just used for interpolating Ri_bulk to determine the OBL_depth).

Then, when I wrote the routine to evaluate a cubic polynomial, I assumed it was of the form

f(x) = a(1) + a(2)*x + a(3)*x^2 + a(4)*x^3

(This was just used the shape function.)

However, for computing G(1) and G'(1), I need to interpolate nu AND evaluate the interpolating polynomial at OBL_depth, so these routines need to order the coefficients the same way. I chose to go with the standard set in evaluate_cubic because
(a) It allows the shorthand f(x) = sum a(n)*x^(n-1)
(b) I _think_ the interpolation routine / root-finder is only used internally, so switching to this convention won't require any re-coding in POP, MPAS, or MOM (I didn't change anything in the stand-alone driver).

This issue cropped up as I was adding the "compute_nu_at_OBL_depth" routine, hence the block of code that is ifdef-ed out at the bottom of cvmix_kpp.F90. I'll finish writing that routine now that I can compute an interpolant AND evaulate it at a given value.