I _think_ the step to compute the shape function (specifically, G(1) and G'(1)) is set up to allow the user to determine whether to compute G'(1) as outlined in LMD94 or just set G'(1) = 0. Right now, this is done by setting the flag lnoDGat1 in cvmix_init_kpp [by default, the flag is set to .true. and G'(1) = 0].
I don't have a feel for reasonable values for the buoyancy flux or the turbulent friction velocity at the surface, so I don't have a good test to see if this code does what we expect. I'm fairly sure that the sign is correct on all of the terms, though.