A couple of updates for the non-local transport term:
1) As Steve pointed out, the momentum term should have 2 components (x & y) -- they are still both set to zero for now, but I want everything in place in case we want to see the effect of the Smyth et al. non-local momentum terms.

2) These terms now output G_x*Cs, which is only a portion of K_x*gamma_x: based on the notation in Steve's latest draft of the manual, the user is responsible for multiplying this by the [negative] penetrative heat flux or the freshwater flux to reconstruct Eq. (7.6).

This commit also introduces a couple of local variables into cvmix_coeffs_kpp: OBL_diff and OBL_visc. Hopefully it's clear from the names, but OBL_diff is the diffusivity in the OBL and OBL_visc in the viscosity there. I had been over-writing the diff and visc values passed in from the model, but realized that I may need to reference the old values in the enhanced mixing section.

Speaking of which, the "enhanced mixing" portion of KPP is the only thing left to be coded up. Then the real fun (testing / bug-fixing) begins.