Adding routines to compute bulk Richardson number and the unresolved shear (needed for bulk Richardson number) in cvmix_kpp. You can either pass the unresolved shear to the compute_bulk_Richardson number or pass the buoyancy and turbulent velocity scale at each level and the routine will compute the unresolved shear for you.
I also added test #5 to the stand-alone driver, but it still needs some work (and I need some help interpretting the results... I'll send out an email about it after this commit). For starters, I need to add the &kpp_col5_nml namelist to the driver to read in some of the parameters that can be set.

As we talked about on the phone, I'm also coming across many routines where it would be nice to know the units of various inputs and outputs... I'll make a note of that on the issue page.