I got pulled in four different directions this week and finally had a couple hours to try to finish up the stand-alone KPP test so I can move on to POP+CVmix. This update adds output to the KPP driver for the "figure C1" test; the plot_bulk_Rich.ncl script doesn't actually recreate figure C1 yet but it does get the buoyancy figure spot on.
The one change to shared code is to include buoyancy, horizontal velocity (both components), and bulk Richardson number in the CVMix data type wrapper - needed for the output routine. I think this will prove useful for making wrapped interfaces for a lot of the new computation routines that have come into the KPP module (computing unresolved shear or bulk Richardson number, for example).